% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_info.R
\docType{data}
\name{A_germline}
\alias{A_germline}
\title{A matrix of read numbers of alternative alleles}
\format{
a matrix of float
}
\source{
A fibroblast sample from HipSci project
}
\usage{
example_donor
}
\value{
NULL, but makes available a matrix
}
\description{
This matrix contains read numbers of alternative alleles for
34 germline variants (near the somatic variants) across 428 cells,
from one example scRNA-seq sample
}
\author{
Yuanhua Huang, Davis McCarthy, 2018-06-25
}
