% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment.R
\name{assign_curve}
\alias{assign_curve}
\title{Calculate assignability and accuracy along prob_gap}
\usage{
assign_curve(prob_mat, simu_mat)
}
\arguments{
\item{prob_mat}{A matrix (M cells x K clones), estiamted probability of cell 
to clone}

\item{simu_mat}{A matrix (M cells x K clones), tru cell-clone assignment}
}
\description{
Calculate assignability and accuracy along prob_gap
}
