% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment.R
\name{assign_macro_ROC}
\alias{assign_macro_ROC}
\title{Calculate macro ROC}
\usage{
assign_macro_ROC(prob_mat, simu_mat)
}
\arguments{
\item{prob_mat}{A matrix (M cells x K clones), estiamted probability of cell 
to clone}

\item{simu_mat}{A matrix (M cells x K clones), tru cell-clone assignment}
}
\description{
Calculate macro ROC
}
