% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_assign_EM.R
\name{cell_assign_EM}
\alias{cell_assign_EM}
\title{EM algorithm for assigning cells to clones and estimating parameters in 
mixture model.
The two Bernoulli components correspond to false positive and false negative.
The two binomial components correspond to reads distribution without and with
variant.}
\usage{
cell_assign_EM(A, D, C, Psi = NULL, model = "Bernoulli", threshold = 1,
  threshold_type = "count", max_iter = 1000, min_iter = 10,
  logLik_threshold = 1e-05, verbose = TRUE)
}
\arguments{
\item{A}{A matrix of integers. Number of alteration reads in variant i cell j}

\item{D}{A matrix of integers. Number of reads depth in variant i cell j}

\item{C}{A matrix of binary values. The clone-variant configuration, whcih 
encodes the phylogenetic tree structure. This is the output Z of Canopy}

\item{Psi}{A vector of float. The fractions of each clone, output P of Canopy}

\item{model}{A string. The model to use: Bernoulli or binomial}

\item{threshold}{A value of integer or float. the threshold on count or 
fraction of alteration reads when using Bernoulli model}

\item{threshold_type}{A string. The type of threshold: count or fraction}

\item{max_iter}{A integer. The maximum number of iterations in EM algorithm. 
The real iteration may finish earlier.}

\item{min_iter}{A integer. The minimum number of iterations in EM algorithm.}

\item{logLik_threshold}{A float. The threshold of logLikelihood increase for
detecting convergence.}

\item{verbose}{logical(1), should the function output verbose information as
it runs?}
}
\value{
a list containing \code{alpha}, a float denoting the estimated false 
positive rate, \code{beta}, a float denoting the estimated false negative 
rate, \code{prob}, the matrix of fitted probabilities of each cell belonging 
to each clone, and \code{logLik},
the log likelihood of the parameter based on the final cell assignment.

a list containing \code{theta}, a vector of two floats denoting the 
parameters of the two componets of the base model, i.e., mean of Bernoulli or 
binomial model given variant exists or not, \code{prob}, the matrix of 
posterior probabilities of each cell belonging to each clone with fitted 
parameters, and \code{logLik}, the log likelihood of the final parameters.
}
\description{
EM algorithm for assigning cells to clones and estimating parameters in 
mixture model.
The two Bernoulli components correspond to false positive and false negative.
The two binomial components correspond to reads distribution without and with
variant.
}
