% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment.R
\name{colMatch}
\alias{colMatch}
\title{Column match between two matrices by the minimum mean abosolute difference}
\usage{
colMatch(A, B)
}
\arguments{
\item{A}{The first matrix which will be matched}

\item{B}{The second matrix, the return index will be used on}
}
\value{
\code{idx}, the column index of B to be matched to A
}
\description{
Column match between two matrices by the minimum mean abosolute difference
}
