% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_info.R
\docType{data}
\name{Config_all}
\alias{Config_all}
\title{A list of tree configuration}
\format{
a list of list of matrix
}
\source{
PASTRI Python package
}
\usage{
config_all
}
\value{
NULL, but makes available a list
}
\description{
This list of tree configuration between 3 clones to 10 clones,
each element is a list with all possible tree matrix
}
\author{
Yuanhua Huang, Davis McCarthy, 2018-06-25
}
