% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmaps.R
\name{confusion_heatmap}
\alias{confusion_heatmap}
\title{Plot confusion heatmap for assessing simulation accuracy}
\usage{
confusion_heatmap(prob_mat, sim_mat, prob_gap_min = 0, pre_title = "")
}
\arguments{
\item{prob_mat}{A matrix (M x K), the estimated probability of cell j to 
clone k}

\item{prob_gap_min}{A float value, only show confusion matrix on cells with 
prob_gap >= prob_gap_min}

\item{prob_mat}{A matrix (M x K), the true assignment of cell to clone}
}
\description{
Plot confusion heatmap for assessing simulation accuracy
}
