% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_info.R
\docType{data}
\name{D_clone}
\alias{D_clone}
\title{A matrix of sequencing depths for clone ID}
\format{
a matrix of float
}
\source{
A fibroblast sample from HipSci project
}
\usage{
example_donor
}
\value{
NULL, but makes aavailable a matrix
}
\description{
This matrix contains sequncing depths for 34 somatic variants
across 428 cells, from one example scRNA-seq sample
}
\author{
Yuanhua Huang, Davis McCarthy, 2018-06-25
}
