% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_info.R
\docType{data}
\name{D_germline}
\alias{D_germline}
\title{A matrix of sequencing depths}
\format{
a matrix of float
}
\source{
A fibroblast sample from HipSci project
}
\usage{
example_donor
}
\value{
NULL, but makes available a matrix
}
\description{
This matrix contains sequencing depths for 34 germline variants
(near the somatic variants) across 428 cells, from one example scRNA-seq
sample
}
\author{
Yuanhua Huang, Davis McCarthy, 2018-06-25
}
