% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_info.R
\docType{data}
\name{D_input}
\alias{D_input}
\title{A matrix of sequncing depths}
\format{a matrix of float}
\source{
A fibroblast sample from HipSci project
}
\usage{
simulation_input
}
\value{
NULL, but makes aavailable a matrix
}
\description{
This matrix contains sequncing depths for 439 somatic variants 
across 151 cells, from one particular scRNA-seq sample, can be used to 
generate sequncing depths
}
\author{
Yuanhua Huang, Davis McCarthy, 2018-06-25
}
