% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clone_id.R
\name{devianceIC}
\alias{devianceIC}
\title{Deviance Information Criterion for cardelino model}
\usage{
devianceIC(logLik_all, logLik_post)
}
\arguments{
\item{logLik_all}{A vector of numeric; the log likelihood of posterior
sample, i.e., posterior samples of deviance}

\item{logLik_post}{numeric(1); the log likelihood of mean posterior
parameters, i.e., deviance of posterior means}
}
\value{
\code{DIC}, a float of deviance information criterion
}
\description{
Deviance Information Criterion for cardelino model
}
\author{
Yuanhua Huang
}
