% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donor_id.R
\name{donor_id}
\alias{donor_id}
\title{Donor deconvolution of scRNA-seq data, replaced by function vireo}
\usage{
donor_id(...)
}
\description{
Donor deconvolution of scRNA-seq data, replaced by function vireo
}
