% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donor_id.R
\name{donor_id_EM}
\alias{donor_id_EM}
\title{EM algorithm for donor deconvolution with or without genotypes.}
\usage{
donor_id_EM(A, D, GT = NULL, K = NULL, gt_singlet = c(0, 1, 2),
  check_doublet = TRUE, min_iter = 10, max_iter = 200,
  n_init = NULL, logLik_threshold = 1e-05, verbose = TRUE)
}
\arguments{
\item{A}{A matrix of integers. Number of alteration reads in SNP i cell j}

\item{D}{A matrix of integers. Number of reads depth in SNP i cell j}

\item{GT}{A matrix of integers for genotypes. The donor-SNP configuration.}

\item{K}{An integer. The number of donors to infer if not given GT.}

\item{gt_singlet}{A vector of integers. The candidate elements of GT. Default
is c(0, 1, 2); but can be others. Note, gt will be averaged for doublet.
When GT is given, gt_singlet will be adapted to GT.}

\item{check_doublet}{logical(1), if TRUE, check doublet, otherwise ignore.}

\item{min_iter}{A integer. The minimum number of iterations in EM algorithm.}

\item{max_iter}{A integer. The maximum number of iterations in EM algorithm.
The real iteration may finish earlier.}

\item{n_init}{A integer. The number of random initializations for EM
algorithm, which can be useful to avoid local optima if not given genotypes.
Default: 1 if given GT, 5 if not given GT.}

\item{logLik_threshold}{A float. The threshold of logLikelihood increase for
detecting convergence.}

\item{verbose}{logical(1), If TRUE, output verbose information when running.}
}
\value{
a list containing
\code{logLik}, the log likelihood.
\code{theta}, a vector denoting the binomial parameters for each genotype.
\code{prob}, a matrix of posterior probability of cell assignment to donors.
The summary may less than 1, as there are some probabilities go to doublets.
\code{prob_doublet}, a matrix of posterior probability of cell assignment to
each inter-donor doublet.
\code{GT}, the input GT or a point estimate of genotype of donors. Note,
this may be not very accurate, especially for lowly expressed SNPs.
\code{GT_doublet}, the pair-wise doublet genotype based on GT.
}
\description{
EM algorithm for donor deconvolution with or without genotypes.
}
\details{
Users should typically use \code{\link{donor_id}} rather than this
lower-level function.
}
\examples{
data(example_donor)
res <- donor_id_EM(A_clone, D_clone, GT = tree$Z[1:nrow(A_clone),])
head(res$prob)

}
