% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donor_id.R
\name{donor_id_GT}
\alias{donor_id_GT}
\title{EM algorithm for donor deconvolution in multiplexed scRNA-seq with genotypes.}
\usage{
donor_id_GT(A, D, GT, check_doublet = TRUE, n_vars_threshold = 10,
  p_threshold = 0.95, Psi = NULL, min_iter = 10, max_iter = 1000,
  logLik_threshold = 1e-05, verbose = TRUE)
}
\arguments{
\item{A}{A matrix of integers. Number of alteration reads in SNP i cell j}

\item{D}{A matrix of integers. Number of reads depth in SNP i cell j}

\item{GT}{A matrix of integers for genotypes. The donor-SNP configuration,
the element should be 0, 1, 2, but others are probably also appliable.}

\item{check_doublet}{logical(1), if TRUE, check doublet, otherwise ignore.}

\item{n_vars_threshold}{integer(1), if the number of variants with coverage
in a cell is below this threshold, then the cell will be given an
"unassigned" donor ID (default: 10)}

\item{p_threshold}{numeric(1), threshold for posterior probability of
donor assignment (must be in [0, 1]); if best posterior probability for a
donor is greater then the threshold, then the cell is assigned to that donor
(as long as the cell is not determined to be a doublet) and if below the
threshold, then the cell's donor ID is "unassigned"}

\item{Psi}{A vector of float. The fractions of each donor; default uniform.}

\item{min_iter}{A integer. The minimum number of iterations in EM algorithm.}

\item{max_iter}{A integer. The maximum number of iterations in EM algorithm.
The real iteration may finish earlier.}

\item{logLik_threshold}{A float. The threshold of logLikelihood increase for
detecting convergence.}

\item{verbose}{logical(1), If TRUE, output verbose information when running.}
}
\value{
a list containing
\code{logLik}, the log likelihood.
\code{theta}, a vector denoting the binomial parameters for each genotype.
\code{prob}, a matrix of posterior probability of cell assignment to donors.
The summary may less than 1, as there are some probabilities go to doublets.
\code{prob_doublet}, a matrix of posterior probability of cell assignment to
each inter-donor doublet.
}
\description{
EM algorithm for donor deconvolution in multiplexed scRNA-seq with genotypes.
}
\details{
Users should typically use \code{\link{donor_id}} rather than this
lower-level function.
}
\examples{
data(example_donor)
ids <- donor_id_GT(A, D, GT = tree$Z[1:nrow(A),])
table(ids$assigned$donor_id)

}
