% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donor_id.R
\name{donor_id_VB}
\alias{donor_id_VB}
\title{Variational inference for donor deconvolution with or without genotypes.}
\usage{
donor_id_VB(A, D, K = NULL, K_amplify = 1.5, GT = NULL,
  GT_prior = NULL, n_init = NULL, n_proc = 4, random_seed = NULL,
  ...)
}
\arguments{
\item{A}{A matrix of integers. Number of alteration reads in SNP i cell j}

\item{D}{A matrix of integers. Number of reads depth in SNP i cell j}

\item{K}{An integer. The number of donors to infer if GT is not given nor 
complete.}

\item{K_amplify}{A float. The amplify ratio of donor number in the first run}

\item{GT}{A matrix of integers for genotypes. The donor-SNP configuration.}

\item{n_proc}{An integer. The number of processors to use.}

\item{...}{arguments passed to \code{run_VB}}
}
\value{
a list containing
\code{logLik}, the log likelihood.
\code{theta}, a vector denoting the binomial parameters for each genotype.
\code{prob}, a matrix of posterior probability of cell assignment to donors.
The summary may less than 1, as there are some probabilities go to doublets.
\code{prob_doublet}, a matrix of posterior probability of cell assignment to
each inter-donor doublet.
\code{GT}, the input GT or a point estimate of genotype of donors. Note,
this may be not very accurate, especially for lowly expressed SNPs.
\code{GT_doublet}, the pair-wise doublet genotype based on GT.
}
\description{
Variational inference for donor deconvolution with or without genotypes.
}
\details{
Users should typically use \code{\link{donor_id}} rather than this
lower-level function.
}
\examples{
data(example_donor)
res <- donor_id_VB(A_clone, D_clone, GT = tree$Z[1:nrow(A_clone),])
head(res$prob)

}
