% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donor_id.R
\name{get_doublet_GT}
\alias{get_doublet_GT}
\title{Generate genotype probability for doublets}
\usage{
get_doublet_GT(GT_prob, K)
}
\arguments{
\item{GT_prob}{A matrix of genotype for singlets}

\item{K}{An integer for number of donors}
}
\value{
\code{GT_both}, a matrix of genotype probability for both singlet
and doublet donors
}
\description{
Generate genotype probability for doublets
}
