% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donor_id.R
\name{get_doublet_theta}
\alias{get_doublet_theta}
\title{Generate theta parameters for doublet genotype}
\usage{
get_doublet_theta(theta_shapes)
}
\arguments{
\item{theta_shapes}{A 3-by-2 matrix of beta paramters for genotype 0, 1, 2}
}
\value{
a 5-by-2 matrix of beta paramters for genotype 0, 1, 2, 0.5, 1.5
}
\description{
Generate theta parameters for doublet genotype
}
