% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_assign_Gibbs.R
\name{get_Gibbs_prob}
\alias{get_Gibbs_prob}
\title{Get the probability matrix from Gibbs sampling on assignment}
\usage{
get_Gibbs_prob(assign_Gibbs, buin_in = 0.25)
}
\arguments{
\item{assign_Gibbs}{A matrix of T iterations on M cells}

\item{buin_in}{A float of the fraction of buin-in iterations}
}
\value{
A probability matrix of M cells and K clones
}
\description{
Get the probability matrix from Gibbs sampling on assignment
}
