% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donor_id.R
\name{get_ID_prob}
\alias{get_ID_prob}
\title{Internal function to update cell assignement probability}
\usage{
get_ID_prob(A, D, GT_prob, theta_shapes, Psi)
}
\arguments{
\item{A}{A matrix of integers. Number of alteration reads in SNP i cell j}

\item{D}{A matrix of integers. Number of reads depth in SNP i cell j}

\item{GT_prob}{A matix of float, with the size of N*K-by-3 if there are 3 
genotypes.}

\item{theta_shapes}{A matrix of float with size 3-by-2. Each row is the beta
prior distribution parameters of the according genotype}

\item{Psi}{A voctor of float. The fraction of each donor.}
}
\value{
A list containing \code{logLik} and \code{ID_prob}
}
\description{
Internal function to update cell assignement probability
}
