% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clone_id.R
\name{get_logLik}
\alias{get_logLik}
\title{Log likelihood of clone_id model
It returns P(A, D | C, I, theta0, theta1)}
\usage{
get_logLik(A1, B1, Config, Assign, theta0, theta1)
}
\arguments{
\item{A1}{variant x cell matrix of integers; number of alternative allele
reads in variant i cell j}

\item{B1}{variant x cell matrix of integers; number of reference allele
reads in variant i cell j}

\item{Config}{variant x clone matrix of float values. The clone-variant
configuration probability, averaged by posterior samples}

\item{Assign}{cells x clone matrix of float values. The cell-clone
assignment probability, averaged by posterior samples}

\item{theta0}{the binomial rate for alternative allele from config = 0}

\item{theta1}{the binomial rate for alternative allele from config = 1}
}
\value{
\code{logLik}, a float of log likelihood
}
\description{
Log likelihood of clone_id model
It returns P(A, D | C, I, theta0, theta1)
}
\author{
Yuanhua Huang
}
