% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_utils.R
\name{get_prob_gap}
\alias{get_prob_gap}
\title{Get the probability gap between the highest and the second highest assignmnet
probabilities.}
\usage{
get_prob_gap(prob_assign)
}
\arguments{
\item{prob_assign}{A matrix of floats. Clonal assignment probability of N 
cells to K clones.}
}
\description{
Get the probability gap between the highest and the second highest assignmnet
probabilities.
}
