% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_utils.R
\name{get_prob_label}
\alias{get_prob_label}
\title{Get the clone label from the assignmnet probabilities.}
\usage{
get_prob_label(prob_assign)
}
\arguments{
\item{prob_assign}{A matrix of floats. Clonal assignment probability of N 
cells to K clones.}
}
\value{
a vector of the index of clone for each cell. Note, when multiple 
clones have the same assignment probability, only the earliest clone will be 
return. Normally, these cells should be filtered for analysis.
}
\description{
Get the clone label from the assignmnet probabilities.
}
\examples{
data(example_donor)
assignments <- clone_id(A, D, C = tree$Z, inference = "EM")
clone_idx <- get_prob_label(assignments$prob)

}
