% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_utils.R
\name{get_prob_value}
\alias{get_prob_value}
\title{Get the collapsed probability value}
\usage{
get_prob_value(prob_assign, mode = "best")
}
\arguments{
\item{prob_assign}{A matrix of floats. Clonal assignment probability of M
cells to K clones.}

\item{mode}{A string, the mothod for defining scores for filtering cells:
best, second and delta. \code{best}: highest probability of a cell to K
clones, similarly for the \code{second}. \code{delta} is the difference
between the best and the second.}
}
\value{
a vector of the collapsed probability value for each cell, depending
on the mode used.
}
\description{
Get the collapsed probability value
}
\examples{
data(example_donor)
assignments <- clone_id(A_clone, D_clone, Config = tree$Z, inference = "EM")
prob_val <- get_prob_value(assignments$prob, mode = "best")

}
