% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_utils.R
\name{get_snv_theta}
\alias{get_snv_theta}
\title{Calculate the SNV specific allelic frequency (True postive rate) and error
rate (False possitive rate)}
\usage{
get_snv_theta(C, D, A, prob_mat, threshold = 1)
}
\description{
Calculate the SNV specific allelic frequency (True postive rate) and error
rate (False possitive rate)
}
