% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clone_id.R
\name{Geweke_Z}
\alias{Geweke_Z}
\title{Geweke diagnostic for MCMC sampling.}
\usage{
Geweke_Z(X, first = 0.1, last = 0.5)
}
\arguments{
\item{X}{A matrix of MCMC samples for N samples per K variables}

\item{first}{A float between 0 and 1. The initial region of MCMC chain.}

\item{last}{A float between 0 and 1. The final region of MCMC chain.}
}
\value{
\code{Z}, a vector of absolute value of Z scores for each variable.
}
\description{
Geweke diagnostic for MCMC sampling.
}
