% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donor_id.R
\name{GT_to_prob}
\alias{GT_to_prob}
\title{Convert genotype matrix to genotype probability matrix}
\usage{
GT_to_prob(GT, gt_singlet = c(0, 1, 2))
}
\arguments{
\item{GT}{A matrix of genotype}

\item{gt_singlet}{A list of singlet genotyoe}
}
\description{
Convert genotype matrix to genotype probability matrix
}
