% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmaps.R
\name{heatmap.theme}
\alias{heatmap.theme}
\title{The theme of heatmaps for prob_heatmap and sites_heatmap}
\usage{
heatmap.theme(legend.position = "bottom", size = 12)
}
\description{
The theme of heatmaps for prob_heatmap and sites_heatmap
}
