% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{load_cellSNP_vcf}
\alias{load_cellSNP_vcf}
\title{Load sparse matrices A and D from cellSNP VCF file with filtering SNPs}
\usage{
load_cellSNP_vcf(vcf_file, max_other_allele = 0.05, min_count = 20,
  min_MAF = 0.1)
}
\arguments{
\item{vcf_file}{character(1), path to VCF file generated from cellSNP}

\item{max_other_allele}{maximum ratio of other alleles comparing to REF and 
ALT alleles}

\item{min_count}{minimum count across all cells}

\item{min_MAF}{minimum minor allele fraction}
}
\description{
Load sparse matrices A and D from cellSNP VCF file with filtering SNPs
}
\examples{

}
