% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{load_vcf_h5}
\alias{load_vcf_h5}
\title{Load sparse matrices A and D from cellSNP VCF file in HDF5 format}
\usage{
load_vcf_h5(filename)
}
\arguments{
\item{filename}{character(1), path to HDF5 file generated from cellSNP}
}
\description{
Load sparse matrices A and D from cellSNP VCF file in HDF5 format
}
