% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reads_simulator.R
\name{missing_update}
\alias{missing_update}
\title{Update matrix D with manully selected missing rate}
\usage{
missing_update(D, missing_rate = NULL)
}
\description{
Update matrix D with manully selected missing rate
}
