% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment.R
\name{mtx_to_df}
\alias{mtx_to_df}
\title{Convert a matrix to data frame}
\usage{
mtx_to_df(X)
}
\arguments{
\item{X}{A matrix of values}
}
\value{
A data.frame version of the passed matrix.
}
\description{
Convert a matrix to data frame
}
\examples{
mtx_to_df(matrix(seq(12), nrow = 3))
}
