% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment.R
\name{multiPRC}
\alias{multiPRC}
\title{Precision-recall curve for multi-class prediction}
\usage{
multiPRC(
  prob_mat,
  simu_mat,
  marginal_mode = "best",
  cutoff = NULL,
  multiLabel.rm = TRUE,
  add_cut1 = FALSE
)
}
\arguments{
\item{prob_mat}{Probability matrix for each cell to each component}

\item{simu_mat}{The true identity of assignment from simulation}

\item{marginal_mode}{A string for the mode to marginalize the column: best,
second, or delta}

\item{cutoff}{A list of cutoff; if NULL use all unique scores}

\item{multiLabel.rm}{Logical value; if True, remove the samples with
multiple labels}

\item{add_cut1}{Logical value; if True, manually add a cutoff of 1}
}
\description{
Precision-recall curve for multi-class prediction
}
