% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donor_id.R
\name{nega_beta_entropy}
\alias{nega_beta_entropy}
\title{Negative entropy value for beta distribution}
\usage{
nega_beta_entropy(theta_shapes, theta_prior = NULL)
}
\arguments{
\item{theta_shapes}{A matrix of float with size T-by-2. Each row has 
beta parameters [shape1, shape2].}

\item{theta_prior}{A matrix of float with size T-by-2. Each row has 
beta parameters [shape1, shape2]. If theta_prior is NULL, then the entropy
is calculated based on theta_shapes itself, otherwise based on theta_prior.}
}
\value{
A list of T negative entropy values.
}
\description{
Negative entropy value for beta distribution
}
