% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{parse_cell_vcf}
\alias{parse_cell_vcf}
\title{Parse with single-cell variant data from VCF file for cardelino}
\usage{
parse_cell_vcf(vcf_file, filter_variants = TRUE, filter_cells = FALSE,
  cell_nvars_threshold = 1.5, verbose = TRUE, ...)
}
\arguments{
\item{vcf_file}{path to VCF file from which to read data}

\item{filter_variants}{logical(1), should variants that do not have any read
coverage in any cell be filtered out?}

\item{filter_cells}{logical(1), should cells with fewer than a threshold
number of variants with read coverage be filtered out?}

\item{cell_nvars_threshold}{numeric(1), threshold for the number of variants
with non-zero read coverage; if \code{filter_cells} is \code{TRUE} then cells
with a number of variants with read coverage below this threshold will be
filtered out.}

\item{verbose}{logical(1), should messages be printed as function runs?}

\item{...}{further arguments passed to \code{\link[vcfR]{read.vcfR}}}
}
\value{
a list with elements: \code{A}, a variant x cell matrix of read
counts supporting the alternative allele; \code{D}, a variant x cell matrix
of total read depth for each variant (set to \code{NA} if depth is zero);
\code{R}, a variant x cell matrix of the read counts supporting the reference
allele.
}
\description{
Parse with single-cell variant data from VCF file for cardelino
}
\examples{
vcf <- system.file("extdata", "cell_example.mpileup.vcf.gz",
                   package = "cardelino")
input_data <- parse_cell_vcf(vcf, filter_variants = TRUE)

}
\author{
Davis McCarthy
}
