% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmaps.R
\name{prob_heatmap}
\alias{prob_heatmap}
\title{Plot a heatmap for probability of clone assignment}
\usage{
prob_heatmap(prob_mat, prob_gap_min = 0.2, cell_idx = NULL)
}
\arguments{
\item{prob_mat}{A matrix (M x K), the probability of cell j to clone k}

\item{prob_gap_min}{A float, the threshold for assignable cells, i.e., the 
minimum probability gap between the highest and the second clones. It is used
for calculating the fraction of clone}

\item{cell_idx}{A vector the indices of the input cells. If NULL, order by 
the probabilty of each clone}
}
\description{
Plot a heatmap for probability of clone assignment
}
