% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment.R
\name{rowArgmax}
\alias{rowArgmax}
\title{Column index of the maximum value for each row in a matrix}
\usage{
rowArgmax(X)
}
\arguments{
\item{X}{A matrix of floats.}
}
\value{
a vector of the index of column for each row. Note, when multiple
columns have the same value, only the earliest column will be
returned.
}
\description{
Column index of the maximum value for each row in a matrix
}
