% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment.R
\name{rowMax}
\alias{rowMax}
\title{Maximum value for each row in a matrix}
\usage{
rowMax(X, mode = "best")
}
\arguments{
\item{X}{A matrix of floats.}

\item{mode}{A string, the mothod for defining scores for filtering cells:
best, second and delta. \code{best}: highest value for each row, similarly 
for the \code{second}. \code{delta} is the difference between the best and 
the second.}
}
\value{
a vector of the collapsed value for each row, depending
on the mode used.
}
\description{
Maximum value for each row in a matrix
}
