% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reads_simulator.R
\name{sample_tree_SNV}
\alias{sample_tree_SNV}
\title{Down sample number of SNVs in the tree}
\usage{
sample_tree_SNV(tree, n_SNV = NULL)
}
\arguments{
\item{tree}{A tree object from Canopy}

\item{n_SNV}{A integer, the number of SNVs to keep in the output tree}
}
\value{
a phylo tree with down sampled variants
}
\description{
Down sample number of SNVs in the tree
}
\examples{
data(simulation_input)
tree_lite <- sample_tree_SNV(tree_4clone, n_SNV = 10)
}
