% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reads_simulator.R
\name{sim_read_count}
\alias{sim_read_count}
\title{Sythetic reads generator for genetic variants}
\usage{
sim_read_count(C, D, Psi = NULL, means = c(0.03, 0.45), vars = c(0.1,
  0.1), missing_rate = NULL, cell_num = 300)
}
\arguments{
\item{C}{A matrix of binary values. The clone-variant configuration, whcih 
encodes the phylogenetic tree structure. This is the output Z of Canopy}

\item{D}{A vector of integers. Sequencing depth for variants}

\item{Psi}{A vector of float. The fractions of each clone, output P of Canopy}
}
\description{
There are following steps to generate the simulated reads counts for variants
in single cells:
1) given the clonal genotype and the clonal prevalence, the genotypes of 
cells will be generated following a multinormial distribution. 
2) given the distribution of reads coverage, e.g., a set of read coverage 
from real data, the total reads of each variant will be generated by random 
sampling.
3) the allelic frequence of each variant will be generated by following a 
beta distribution.
4) Given the genotype of a cell, if the mutation exists in a cell, the 
alteration read counts will be generated by a binomial distribution, 
papramitized the allelic frequency, sampled from step 3.
5) Given the genotype of a cell, if the mutation doesnot exist in a cell, 
the alteration read counts will be genenrated by a binomial distribution,
paramerized by the technical error rate.
6) Given the missing rate, NA value will be given to mask the read counts
}
