% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmaps.R
\name{sites_heatmap}
\alias{sites_heatmap}
\title{Plot a heatmap for number of mutation sites in each cell}
\usage{
sites_heatmap(Config, A, prob_mat, mode = "present")
}
\arguments{
\item{Config}{A matrix (N x K), clonal genotype configuration}

\item{A}{A matrix (N x M), the present of alternative reads. NA means missing}

\item{prob_mat}{A matrix (M x K), the probability of cell j to clone k}

\item{mode}{A string: present or absent}
}
\description{
Plot a heatmap for number of mutation sites in each cell
}
