% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donor_id.R
\name{transpose_GT_prob}
\alias{transpose_GT_prob}
\title{Transpose genotype probability matrix}
\usage{
transpose_GT_prob(GT_prob, K)
}
\arguments{
\item{GT_prob}{A N*K-by-T matrix of genotype probability}

\item{K}{An integer}
}
\value{
A N*T-by-K matrix of genotype probability
}
\description{
Transpose genotype probability matrix
}
