% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_info.R
\docType{data}
\name{tree}
\alias{tree}
\title{A tree object}
\format{a tree object}
\source{
A fibroblast sample from HipSci project
}
\usage{
example_donor
}
\value{
NULL, but makes aavailable a tree object
}
\description{
This tree object contains clonal tree information, inferred 
from bulk exome-seq data
}
\author{
Yuanhua Huang, Davis McCarthy, 2018-06-25
}
