% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_info.R
\docType{data}
\name{tree_3clone}
\alias{tree_3clone}
\title{A tree object}
\format{
a tree object
}
\source{
A fibroblast sample from HipSci project
}
\usage{
simulation_input
}
\value{
NULL, but makes available a tree object
}
\description{
This tree object with 3 clones contains clonal tree information,
inferred from bulk exome-seq data
}
\author{
Yuanhua Huang, Davis McCarthy, 2018-06-25
}
