% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reads_simulator.R
\name{assign_score}
\alias{assign_score}
\title{Assess performance with simulated data}
\usage{
assign_score(prob_mat, simulate_mat, prob_gap_min = 0.2)
}
\arguments{
\item{prob_mat}{A matrix (M cells x K clones), estiamted probability of cell 
to clone}

\item{simulate_mat}{A matrix (M cells x K clones), tru cell-clone assignment}

\item{prob_gap_min}{A float value, the threshold for assignable cells}
}
\description{
Return fraction of assignable cells, accuracy with assignable cells, and
overall accuracy
}
