% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment.R
\name{assign_scores}
\alias{assign_scores}
\title{Scoring the simulation in assignment of singlets and doublets}
\usage{
assign_scores(prob, I_sim, cutoff = seq(0, 1, 0.001))
}
\arguments{
\item{prob}{Probability matrix for each cell to each component}

\item{I_sim}{The true identity of assignment from simulation}

\item{cutoff}{A list of cutoff from 0 to 1}
}
\description{
Scoring the simulation in assignment of singlets and doublets
}
