% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_assign_Gibbs.R
\name{cell_assign_Gibbs}
\alias{cell_assign_Gibbs}
\title{Gibbs sampling the cell assignments to cloens and the parameters of the
mixture model.
The two Bernoulli components correspond to false positive and false negative.
The two binomial components correspond to reads distribution without and with
variant.}
\usage{
cell_assign_Gibbs(A, D, C, Psi = NULL, A_germ = NULL, D_germ = NULL,
  prior0 = c(0.3, 29.7), prior1 = c(2.25, 2.65), model = "binomial",
  Bern_threshold = 1, max_iter = 10000, min_iter = 1000,
  wise = "variant", verbose = TRUE)
}
\arguments{
\item{A}{A matrix of integers. Number of alteration reads in variant i cell j}

\item{D}{A matrix of integers. Number of reads depth in variant i cell j}

\item{C}{A matrix of binary values. The clone-variant configuration, whcih 
encodes the phylogenetic tree structure. This is the output Z of Canopy}

\item{Psi}{A vector of float. The fractions of each clone, output P of Canopy}

\item{A_germ}{A matrix of integers. Number of alteration reads in germline 
heterozygous site for variant i cell j}

\item{D_germ}{A matrix of integers. Number of reads depth in germline 
heterozygous site for variant i cell j}

\item{prior0}{numeric(2), alpha and beta parameters for the Beta prior 
distribution on the inferred false positive rate.}

\item{prior1}{numeric(2), alpha and beta parameters for the Beta prior 
distribution on the inferred (1 - false negative) rate.}

\item{model}{A string. The model to use: Bernoulli or binomial}

\item{Bern_threshold}{An integer. the count threshold of alteration reads 
when using Bernoulli model}

\item{max_iter}{A integer. The maximum number of iterations in the Gibbs 
sampling, even haven't passed the convergence diagnosis}

\item{min_iter}{A integer. The minimum number of iterations in the Gibbs 
sampling. The real iteration may be longer utile the convergence.}

\item{wise}{A string, the wise of parameters for theta1: global, variant, 
element.}
}
\value{
a list containing \code{theta}, a vector of two floats denoting the 
parameters of the two componets of the base model, i.e., mean of Bernoulli or 
binomial model given variant exists or not, \code{assign}, the matrix of 
clonal labels of each cell is assigned, and \code{logLik}, the log likelihood 
of the parameters during sampling. All returns have a length of H, as the 
Gibbs sampling chain.
}
\description{
Gibbs sampling the cell assignments to cloens and the parameters of the
mixture model.
The two Bernoulli components correspond to false positive and false negative.
The two binomial components correspond to reads distribution without and with
variant.
}
\examples{

}
