% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clone_id.R
\name{devianceIC}
\alias{devianceIC}
\title{Deviance Information Criterion for cardelino model}
\usage{
devianceIC(logLik_mean, A1, B1, Config, theta0, theta1, Psi, W_log)
}
\arguments{
\item{logLik_mean}{A float; the log likelihood of a model averaged by 
posterior of parameters}

\item{A1}{variant x cell matrix of integers; number of alternative allele
reads in variant i cell j}

\item{B1}{variant x cell matrix of integers; number of reference allele 
reads in variant i cell j}

\item{Config}{variant x clone matrix of float values. The clone-variant
configuration probability, averaged by posterior samples}

\item{theta0}{the binomial rate for alternative allele from config = 0}

\item{theta1}{the binomial rate for alternative allele from config = 1}

\item{Psi}{A vector of float. The fractions of each clone}

\item{W_log}{A float for the log value of weight}
}
\value{
\code{DIC}, a float of deviance information criterion
}
\description{
Deviance Information Criterion for cardelino model
}
\author{
Yuanhua Huang
}
