% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donor_id.R
\name{donor_id}
\alias{donor_id}
\title{Donor deconvolution of scRNA-seq data}
\usage{
donor_id(cell_vcf, donor_vcf = NULL, n_donor = NULL,
  check_doublet = TRUE, model = "EM", n_vars_threshold = 10,
  p_threshold = 0.95, verbose = TRUE, genome = "GRCh37",
  seq_levels_style = "Ensembl", donors = NULL, ...)
}
\arguments{
\item{cell_vcf}{either character(1), path to a VCF file containing variant
data for cells, or a \code{\link[VariantAnnotation]{CollapsedVCF}} object}

\item{donor_vcf}{either character(1), path to a VCF file containing genotype
data for donors, or a \code{\link[VariantAnnotation]{CollapsedVCF}} object}

\item{n_donor}{integer(1), number of donors to infer if not given genotypes}

\item{check_doublet}{logical(1), should the function check for doublet cells?}

\item{model}{character(1), model for estimating cell donor ids: EM or Gibbs}

\item{n_vars_threshold}{integer(1), if the number of variants with coverage
in a cell is below this threshold, then the cell will be given an
"unassigned" donor ID (default: 10)}

\item{p_threshold}{numeric(1), threshold for posterior probability of
donor assignment (must be in [0, 1]); if best posterior probability for a
donor is greater then the threshold, then the cell is assigned to that donor
(as long as the cell is not determined to be a doublet) and if below the
threshold, then the cell's donor ID is "unassigned"}

\item{verbose}{logical(1), should the function output verbose information
while running?}

\item{genome}{character(1), string indicating the genome build used in the
VCF file(s) (default: "GRCh37")}

\item{seq_levels_style}{character(1), string passed to
\code{\link[GenomeInfoDb]{seqlevelsStyle}} the style to use for
chromosome/contig names (default: "Ensembl")}

\item{donors}{optional character vector providing a set of donors to use, by
subsetting the donors present in the \code{donor_vcf_file}; if \code{NULL}
(default) then all donors present in VCF will be used.}

\item{...}{arguments passed to \code{donor_id_EM} or \code{donor_id_Gibbs}}
}
\value{
a list with elements: \code{logLik}, log-likelihood of the fitted
model; \code{theta}, ; \code{GT}, a matrix of inferred genotypes for each
donor; \code{GT_doublet}, a matrix of inferred genotypes for each possible
doublet (pairwise combinations of donors); \code{prob}, a matrix of posterior
probabilities of donor identities for each cell; \code{prob_doublet}, a
matrix of posterior probabilities for each possible doublet for each cell;
\code{A}, a variant x cell matrix of observed read counts supporting the
alternative allele; \code{D}, a variant x cell matrix of observed read depth;
\code{assigned}, a data.frame reporting the cell-donor assignments with
columns "cell" (cell identifier), "donor_id" (inferred donor, or "doublet" or
"unassigned"), "prob_max" (the maximum posterior probability across donors),
"prob_doublet" (the probability that the cell is a doublet), "n_vars" (the
number of variants with non-zero read depth used for assignment).
}
\description{
Donor deconvolution of scRNA-seq data
}
\details{
This function reads in all elements of the provided VCF file(s) into
memory, so we highly recommend filtering VCFs to the minimal appropriate set
of variants (e.g. with the bcftools software) before applying them to this
function.
}
\examples{
ids <- donor_id(system.file("extdata", "cells.donorid.vcf.gz", package = "cardelino"),
                system.file("extdata", "donors.donorid.vcf.gz", package = "cardelino"))
table(ids$assigned$donor_id)

}
\author{
Davis McCarthy and Yuanhua Huang
}
