% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donor_id.R
\name{donor_id}
\alias{donor_id}
\title{Donor deconvolution of scRNA-seq data}
\usage{
donor_id(cell_data, donor_data = NULL, n_donor = NULL,
  check_doublet = TRUE, n_vars_threshold = 10, s_threshold = 0.9,
  d_threshold = 0.9, verbose = FALSE, ...)
}
\arguments{
\item{cell_data}{either character(1), path to a VCF file containing variant
data for cells, or a list containing A and D matrices}

\item{donor_data}{either character(1), path to a VCF file containing genotype
data for donors, or a matrix for donor genotypes, matched to cell data}

\item{n_donor}{integer(1), number of donors to infer if not given genotypes}

\item{check_doublet}{logical(1), should the function check for doublet cells?}

\item{n_vars_threshold}{integer(1), if the number of variants with coverage
in a cell is below this threshold, then the cell will be given an
"unassigned" donor ID (default: 10)}

\item{s_threshold}{numeric(1), threshold for posterior probability of
donor assignment (must be in [0, 1]); if best posterior probability for a
donor is greater then the threshold, then the cell is assigned to that donor
(as long as the cell is not determined to be a doublet) and if below the
threshold, then the cell's donor ID is "unassigned"}

\item{d_threshold}{numeric(1), threshold for summarised posterior probability
of doublet detection (must be in [0, 1]);}

\item{verbose}{logical(1), should the function output verbose information
while running?}

\item{...}{arguments passed to \code{donor_id_VB}}
}
\value{
a list with elements: \code{logLik}, log-likelihood of the fitted
model; \code{theta}, ; \code{GT}, a matrix of inferred genotypes for each
donor; \code{GT_doublet}, a matrix of inferred genotypes for each possible
doublet (pairwise combinations of donors); \code{prob}, a matrix of posterior
probabilities of donor identities for each cell; \code{prob_doublet}, a
matrix of posterior probabilities for each possible doublet for each cell;
\code{A}, a variant x cell matrix of observed read counts supporting the
alternative allele; \code{D}, a variant x cell matrix of observed read depth;
\code{assigned}, a data.frame reporting the cell-donor assignments with
columns "cell" (cell identifier), "donor_id" (inferred donor, or "doublet" or
"unassigned"), "prob_max" (the maximum posterior probability across donors),
"prob_doublet" (the probability that the cell is a doublet), "n_vars" (the
number of variants with non-zero read depth used for assignment).
}
\description{
Donor deconvolution of scRNA-seq data
}
\details{
This function reads in all elements of the provided VCF file(s) into
memory, so we highly recommend filtering VCFs to the minimal appropriate set
of variants (e.g. with the bcftools software) before applying them to this
function.
}
\examples{
ids <- donor_id(system.file("extdata", "cells.donorid.vcf.gz", package = "cardelino"),
                system.file("extdata", "donors.donorid.vcf.gz", package = "cardelino"))
table(ids$assigned$donor_id)

}
\author{
Davis McCarthy and Yuanhua Huang
}
