% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donor_id.R
\name{donor_id_GT}
\alias{donor_id_GT}
\title{EM algorithm for donor deconvolution in multiplexed scRNA-seq with genotypes.}
\usage{
donor_id_GT(A, D, GT, Psi = NULL, max_iter = 1000, min_iter = 10,
  logLik_threshold = 1e-05, check_doublet = TRUE, verbose = TRUE)
}
\arguments{
\item{A}{A matrix of integers. Number of alteration reads in SNP i cell j}

\item{D}{A matrix of integers. Number of reads depth in SNP i cell j}

\item{GT}{A matrix of integers for genotypes. The donor-SNP configuration, 
the element should be 0, 1, 2, but others are probably also appliable.}

\item{Psi}{A vector of float. The fractions of each donor; default uniform.}

\item{max_iter}{A integer. The maximum number of iterations in EM algorithm. 
The real iteration may finish earlier.}

\item{min_iter}{A integer. The minimum number of iterations in EM algorithm.}

\item{logLik_threshold}{A float. The threshold of logLikelihood increase for
detecting convergence.}

\item{check_doublet}{logical(1), If TRUE, check doublet, otherwise ignore.}

\item{verbose}{logical(1), If TRUE, output verbose information when running.}
}
\value{
a list containing 
\code{logLik}, the log likelihood.
\code{theta}, a vector denoting the binomial parameters for each genotype. 
\code{prob}, a matrix of posterior probability of cell assignment to donors. 
The summary may less than 1, as there are some probabilities go to doublets. 
\code{prob_doublet}, a matrix of posterior probability of cell assignment to 
each inter-donor doublet.
}
\description{
EM algorithm for donor deconvolution in multiplexed scRNA-seq with genotypes.
}
