% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donor_id.R
\name{get_doublet_GT}
\alias{get_doublet_GT}
\title{Generate genotype for doublets}
\usage{
get_doublet_GT(GT)
}
\arguments{
\item{GT}{A matrix of genotype for singlets}
}
\value{
\code{GT_doublet}, a matrix genotype for doublets
}
\description{
Generate genotype for doublets
}
\examples{
GT <- matrix(sample(c(0,1,2), 150, replace = TRUE), nrow = 50)
GT_doublet <- get_doublet_GT(GT)

}
