% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_utils.R
\name{get_prob_value}
\alias{get_prob_value}
\title{Get the collapsed probability value}
\usage{
get_prob_value(prob_assign, mode = "best")
}
\arguments{
\item{prob_assign}{A matrix of floats. Clonal assignment probability of N 
cells to K clones.}

\item{mode}{A string, the mothod for defining scores for filtering cells: 
best, second and delta. Best: highest probability of a cell to K clones, 
similarly for second. delta is the difference between the best and second.}
}
\description{
the collapsed probability value includes 1) the best probability, 2) the
second best probability, and 3) the delta between the best and the second.
}
