% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{heatmap.theme}
\alias{heatmap.theme}
\title{The theme of heatmaps for prob_heatmap and sites_heatmap}
\usage{
heatmap.theme(legend.position = "bottom", size = 12)
}
\arguments{
\item{legend.position}{character, describes where to place legend on plot
(passed to \code{\link[ggplot2]{theme_gray}})}

\item{size}{numeric, base font size for plot (passed to
\code{\link[ggplot2]{theme_gray}})}
}
\value{
a ggplot theme based on theme_gray
}
\description{
The theme of heatmaps for prob_heatmap and sites_heatmap
}
