% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{load_GT_vcf}
\alias{load_GT_vcf}
\title{Load genotype VCF into numeric values: 0, 1, or 2}
\usage{
load_GT_vcf(vcf_file)
}
\arguments{
\item{vcf_file}{character(1), path to VCF file for donor genotypes}
}
\description{
Load genotype VCF into numeric values: 0, 1, or 2
}
