% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmaps.R
\name{vc_heatmap}
\alias{vc_heatmap}
\title{Plot a variant-cell heatmap for cell clonal assignment}
\usage{
vc_heatmap(mat, prob, Config, show_legend = FALSE)
}
\arguments{
\item{mat}{A matrix for heatmap: N variants x M cells. row and column will be 
sorted automatically.}

\item{prob}{A matrix of probability of clonal assignment: M cells x K clones}

\item{Config}{A binary matrix of clonal Configuration: N variants x K clones}

\item{show_legend}{A bool value: if TRUE, show the legend}
}
\value{
a pheatmap object
}
\description{
Plot a variant-cell heatmap for cell clonal assignment
}
\references{
This function makes use of the \code{\link{pheatmap}} packages
}
