% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donor_id.R
\name{vireo_flock}
\alias{vireo_flock}
\title{Variational inference for donor deconvolution with or without genotypes.}
\usage{
vireo_flock(A, D, K = NULL, K_amplify = 1.5, GT = NULL,
  GT_prior = NULL, n_init = NULL, n_proc = 1, random_seed = NULL,
  ...)
}
\arguments{
\item{A}{A matrix of integers. Number of alteration reads in SNP i cell j}

\item{D}{A matrix of integers. Number of reads depth in SNP i cell j}

\item{K}{An integer. The number of donors to infer if GT is not given nor 
complete.}

\item{K_amplify}{A float. The amplify ratio of donor number in the first run}

\item{GT}{A matrix of integers for genotypes. The donor-SNP configuration.}

\item{GT_prior}{A matix of float, with the same size of GT_prob output, i.e.,
N*K-by-3 if there are 3 genotypes.}

\item{n_init}{A integer. The number of random initializations for variational
inference, which can be useful to avoid local optima if not given genotypes.
Default: 1 if given GT, 5 if not given GT.}

\item{n_proc}{An integer. The number of processors to use.}

\item{random_seed}{An integer. The seed for random initialization.}

\item{...}{arguments passed to \code{vireo_core}}
}
\value{
a list containing
\code{logLik}, the log likelihood.
\code{theta}, a vector denoting the binomial parameters for each genotype.
\code{prob}, a matrix of posterior probability of cell assignment to donors.
The summary may less than 1, as there are some probabilities go to doublets.
\code{prob_doublet}, a matrix of posterior probability of cell assignment to
each inter-donor doublet.
\code{GT}, the input GT or a point estimate of genotype of donors. Note,
this may be not very accurate, especially for lowly expressed SNPs.
\code{GT_doublet}, the pair-wise doublet genotype based on GT.
}
\description{
Variational inference for donor deconvolution with or without genotypes.
}
\details{
Users should typically use \code{\link{vireo}} rather than this
lower-level function.
}
